fix: 금액 자료형 int로 변경

+++ src/main/java/site/brainbrain/iqtest/controller/dto/CreateResultRequest.java
@@ -3,6 +3,6 @@ package site.brainbrain.iqtest.controller.dto;
 public record CreateResultRequest(
         String email, String name, int age, String gender, String country,
         String paymentHistory, String answerSheet,
-        String paymentKey, long amount, String orderId
+        String paymentKey, int amount, String orderId
         ) {
 }
