+++ src/main/java/site/brainbrain/iqtest/service/payment/dto/ApiConfirmRequest.java
@@ -2,7 +2,7 @@ package site.brainbrain.iqtest.service.payment.dto;
 
 public record ApiConfirmRequest(
         String paymentKey,
-        long amount,
+        int amount,
         String orderId
 ) {
 }
